Make `test_nonlin.py` pytest-ready
